+++ server/middleware/checkAuth.js
@@ -1,28 +1,56 @@
 const jwt = require('jsonwebtoken')
-const User = require('../models/user')
+// const User = require('../models/user')
 
-module.exports.adminOnly = function(req, res, next){
-  const token = req.body.token || req.headers['token']
-  console.log('Token: ', token)
-  jwt.verify(token, process.env.JWT_PASSPHRASE, (err, decoded) => {
-    if (err) {
-       return res.json({ error: true, msg: 'Invalid Token'})
+module.exports.adminOnly = () => {
+  return function (req, res, next) {
+    if (req.user === null) {
+      return res.status(401).send('Admin only')
+    } else if (!req.user.isAdmin) {
+      return res.status(401).send('Admin only')
     }
-    // TO DO? check if the user exists?
-    if (!decoded.isAdmin) {
-      return res.json({ error: true, msg: 'Admin only access'})
-    }
-    console.log(decoded)
-    req.user = decoded
     next()
-  }) // ends jwt.verity
+  }
 }
 
-module.exports.authUserOnly = (req, res, next) => {
-  const token = req.body.token || req.headers['token']
-  jwt.verify(token, process.env.JWT_PASSPHRASE, (err, decoded) => {
-    if (err) return res.json({ error: true, msg: 'Invalid Token'})
-    req.user = decoded
+module.exports.anyUserOnly = () => {
+  return (req, res, next) => {
+    if (req.user === null) {
+      return res.status(401).send('Logged in Users only')
+    } else if (!req.user) {
+      return res.status(401).send('Logged in Users only')
+    }
     next()
-  })
+  }
+}
+
+module.exports.parseTokenCookie = () => {
+  return (req, res, next) => {
+    // 1. get the token from request cookie
+    const token = req.cookies.token || req.headers['token']
+    // 2. check to see if there is a token ...
+    if (!token) {
+      req.user = null
+      next()
+    } else {
+      // 3. verify the token has not been tampered with ...
+      jwt.verify(token, process.env.JWT_PASSPHRASE, (err, decoded) => {
+        if (err) {
+          req.user = null
+          next()
+        }
+        // 4. valid token save the decoded user data in the request object
+        // so all downstream middleware function have access to user credentials
+        /*
+        ex. of schema of what req.user will look like
+        req.user:  { _id: '58f56dcb6200e636e825383c',
+        username: 'alan',
+        isAdmin: true,
+        exp: 1492526549,
+        iat: 1492522949 }
+        */
+        req.user = decoded
+        next()
+      }) // ends jwt.verity
+    }
+  }
 }