+++ server/public/js/auth_test.js
@@ -1,4 +1,6 @@
-$(document).ready(function(){
+/* global $, axios, localStorage */
+'use strict'
+$(document).ready(function () {
   if (window.location.pathname !== '/') {
     return
   }
@@ -10,40 +12,29 @@ $(document).ready(function(){
   var token = localStorage.getItem('token')
   console.log(token)
 
-  function getAdminData(){
-    axios.get('/api/admin-only/data',{
-      headers: { token: token }
-    }).then(function(response){
-      // console.log(response)
-      if (response.data.error) {
-        console.dir(response.data)
-        var msg = $('<p>').text('You must be an admin to view this data')
-        adminData.append(msg)
-      } else {
-        var msg = $('<p>').text(response.data.msg)
-        adminData.append(msg)
-      }
+  function getAdminData () {
+    let msg
+    axios.get('/api/admin-only/data').then(function (response) {
+      msg = $('<p>').text(response.data.msg)
+      adminData.append(msg)
+    }).catch(function (err) {
+      msg = $('<p>').text('You must be an admin to view this data')
+      adminData.append(msg)
     })
   }
   // console.log('calling getAdmin data ...')
   getAdminData()
 
-  function getUserData(){
-    axios.get('/api/logged-in-only/data', {
-      headers: { token: token }
-    }).then(function(response){
-      if (response.data.error) {
-        console.dir(response.data)
-        var msg = $('<p>').text('You must be signed in to view this data')
-        userData.append(msg)
-      } else {
-        console.log(response.data)
-        var msg = $('<p>').text(response.data.msg)
-        userData.append(msg)
-      }
+  function getUserData () {
+    let msg
+    axios.get('/api/logged-in-only/data').then(function (response) {
+      msg = $('<p>').text(response.data.msg)
+      userData.append(msg)
+    }).catch(function(err) {
+      msg = $('<p>').text('You must be signed in to view this data')
+      userData.append(msg)
     })
   }
   // console.log('calling getUser data ...')
   getUserData()
-
 })