+++ server/server.js
@@ -1,17 +1,31 @@
 const express = require('express')
 const exphbs = require('express-handlebars')
 const bodyParser = require('body-parser')
+const cookieParser = require('cookie-parser')
 const morgan = require('morgan')
 const path = require('path')
+const parseTokenCookie = require('./middleware/checkAuth').parseTokenCookie
 require('dotenv').load()
 
 // ========== Create express app ============
 const app = express()
+app.set('x-powered-by', false)
 const PORT = process.env.PORT || 3000
 app.set('x-powered-by', false)
 
 // ========== middleware ============
 app.use(morgan('dev'))
+app.use(cookieParser())
+app.use(parseTokenCookie()) // custom middleware function I wrote to parse the cookie
+//  ======= testing, cookie and req ==========
+// app.use((req, res, next) => {
+//   console.log('=========COOOKIES=======')
+//   console.log(req.cookies)
+//   console.log('========= END of COOOKIES=======')
+//   next()
+// })
+
+
 app.use(bodyParser.urlencoded({ extended: false }))
 app.use(bodyParser.json())
 app.use(express.static(path.join(__dirname, 'public')))
@@ -24,14 +38,14 @@ app.engine('handlebars', exphbs({
   partialsDir: path.join(__dirname, '/views/partials')
 }))
 
-app.get('/test', (req, res) => { res.json({ msg: 'hello world' })})
+app.get('/test', (req, res) => { res.json({ msg: 'hello world' }) })
 app.use('/', require('./controllers/htmlRouter'))
 app.use('/auth', require('./controllers/authRouter'))
 app.use('/api', require('./controllers/apiRouter'))
 
 // ========== start server ============
 
-if(process.env.NODE_ENV !== 'testing') {
+if (process.env.NODE_ENV !== 'testing') {
   // connect to the database
   require('./models').connect(process.env.MONGODB_URI)
     .then(() => {
@@ -40,8 +54,9 @@ if(process.env.NODE_ENV !== 'testing') {
         console.log(`Listening on port: ${PORT}`)
       })
     })
-    .catch(() => {
+    .catch((err) => {
       console.log('Mongo DB connection error')
+      console.log(err)
     })
 }