+++ CHANGELOG.md
@@ -50,3 +50,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - working Beta version of the JWT authenticated paths without passport!
 - added a demo gif
+
+## [1.0] - 2017-04-17
+### Added
+- added bcrypt for hashing passwords and saving them in mongo db
+
+## [1.1] - 2017-04-18
+### Added
+- Saving the token in a cookie now, so token gets sent to server on every request.
+- refactored client-side js, much simpler and one file
+- refactored auth middleware, so user info gets saved in `req.user` after getting verified
+### BUGS
+- need to clear cache after user logs out